Move /v1 routes in apiBinder.router to v1.ts
