+++ src/compose/application-manager.ts
@@ -1,48 +1,47 @@
 import * as express from 'express';
 import * as _ from 'lodash';
+import { EventEmitter } from 'events';
 import StrictEventEmitter from 'strict-event-emitter-types';
 
 import * as config from '../config';
 import { transaction, Transaction } from '../db';
+import * as logger from '../logger';
+import LocalModeManager from '../local-mode';
+import { Proxyvisor } from '../proxyvisor';
 import * as dbFormat from '../device-state/db-format';
+import { createV2Api } from '../device-api/v2';
+
 import { validateTargetContracts } from '../lib/contracts';
 import constants = require('../lib/constants');
 import { docker } from '../lib/docker-utils';
-import * as logger from '../logger';
 import log from '../lib/supervisor-console';
-import LocalModeManager from '../local-mode';
 import {
 	ContractViolationError,
 	InternalInconsistencyError,
 } from '../lib/errors';
 import { lock } from '../lib/update-lock';
+import { checkTruthy } from '../lib/validation';
 
 import App from './app';
 import * as volumeManager from './volume-manager';
 import * as networkManager from './network-manager';
 import * as serviceManager from './service-manager';
 import * as imageManager from './images';
-import type { Image } from './images';
-import { getExecutors, CompositionStepT } from './composition-steps';
+import { getExecutors, generateStep } from './composition-steps';
 import * as commitStore from './commit';
-
 import Service from './service';
 import Network from './network';
 import Volume from './volume';
 
-import { createV1Api } from '../device-api/v1';
-import { createV2Api } from '../device-api/v2';
-import { CompositionStep, generateStep } from './composition-steps';
-import {
+import type { Image } from './images';
+import type { CompositionStep, CompositionStepT } from './composition-steps';
+import type {
 	InstancedAppState,
 	TargetApps,
 	DeviceLegacyReport,
 	AppState,
 	ServiceState,
 } from '../types/state';
-import { checkTruthy } from '../lib/validation';
-import { Proxyvisor } from '../proxyvisor';
-import { EventEmitter } from 'events';
 
 type ApplicationManagerEventEmitter = StrictEventEmitter<
 	EventEmitter,
@@ -64,7 +63,6 @@ const localModeManager = new LocalModeManager();
 export const router = (() => {
 	const $router = express.Router();
 
-	createV1Api($router);
 	createV2Api($router);
 
 	$router.use(proxyvisor.router);
