+++ src/supervisor.ts
@@ -3,16 +3,18 @@ import * as db from './db';
 import * as config from './config';
 import * as deviceState from './device-state';
 import * as eventTracker from './event-tracker';
+import * as logger from './logger';
+import logMonitor from './logging/monitor';
+import SupervisorAPI from './device-api';
+import * as v1 from './device-api/v1';
+
 import { intialiseContractRequirements } from './lib/contracts';
 import { normaliseLegacyDatabase } from './lib/legacy';
 import * as osRelease from './lib/os-release';
-import * as logger from './logger';
-import SupervisorAPI from './device-api';
 import log from './lib/supervisor-console';
 import version = require('./lib/supervisor-version');
 import * as avahi from './lib/avahi';
 import * as firewall from './lib/firewall';
-import logMonitor from './logging/monitor';
 
 const startupConfigFields: config.ConfigKey[] = [
 	'uuid',
@@ -67,7 +69,7 @@ export class Supervisor {
 
 		log.info('Starting API server');
 		this.api = new SupervisorAPI({
-			routers: [apiBinder.router, deviceState.router],
+			routers: [v1.router, deviceState.router],
 			healthchecks: [apiBinder.healthcheck, deviceState.healthcheck],
 		});
 		this.api.listen(conf.listenPort, conf.apiTimeout);
